Created basic test for shifting heavy hitter work
I added in test 13 and 14 which implements a newly defined function in
the dist_reads.c that utilizes key+ and modulo operation to shift the
output of the heavyhitter generator. Preliminary findings seem to show
that not re-balancing is faster than re-balancing. I will run tests to
see what is causing slower performance in re-balancing.